requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
coloredlogs==15.0.1
distro==1.9.0
durationpy==0.10
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub>=0.34.0,<1.0
humanfriendly==10.0
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy<2.0,>=1.24.0
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
posthog==5.4.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pypika==0.50.0
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers>=0.21,<0.22
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.0
sentence-transformers==2.7.0
# Note: faiss-cpu doesn't support Python 3.13 yet - use Python 3.12
# If you must use Python 3.13, consider using an alternative vector DB or building from source
faiss-cpu==1.8.0.post1
pypdf==5.1.0
python-multipart==0.0.12
email-validator==2.3.0
email-validator==2.2.0